Integration test setup for Octopoes
