+++ .github/workflows/octopoes_itest-crux-regression.yml
@@ -9,7 +9,7 @@ on:
       - octopoes/**
 
 jobs:
-  itest:
+  rtest:
     runs-on: ubuntu-22.04
 
     steps:
@@ -29,7 +29,7 @@ jobs:
         id: buildx
 
       - name: Run integration tests on Crux (regression)
-        run: make itest-crux
+        run: make rtest-crux
         env:
           DOCKER_BUILDKIT: 1
         working-directory: ./octopoes
