+++ octopoes/.ci/ci-multinode.env
@@ -1,0 +1,10 @@
+RABBITMQ_DEFAULT_VHOST=kat
+RABBITMQ_DEFAULT_USER=ci_user
+RABBITMQ_DEFAULT_PASS=ci_pass
+
+QUEUE_URI=amqp://${RABBITMQ_DEFAULT_USER}:${RABBITMQ_DEFAULT_PASS}@rabbitmq:5672/${RABBITMQ_DEFAULT_VHOST}
+
+KATALOGUS_API=http://katalogus_mock:8080
+
+XTDB_URI=http://xtdb_multinode:3000
+XTDB_TYPE=xtdb-multinode
