+++ octopoes/Dockerfile
@@ -18,9 +18,10 @@ ENV PATH=/home/octopoes/.local/bin:${PATH}
 # dependencies
 ARG ENVIRONMENT
 
-COPY requirements.txt .
+COPY requirements.txt requirements-dev.txt .
 RUN --mount=type=cache,target=/root/.cache pip install --upgrade pip \
-    && pip install -r requirements.txt
+    && pip install -r requirements.txt \
+    && if [ "$ENVIRONMENT" = "dev" ]; then pip install -r requirements-dev.txt; fi
 
 USER octopoes
