+++ octopoes/Makefile
@@ -45,13 +45,18 @@ check:
 test:
 	pytest
 
-itest:
+rtest:
 	docker-compose -f docker-compose-base.yml -f .ci/docker-compose.yml up -d --build
 	sleep 1
 	robot -d reports --variablefile tests/robot/variables.py:xtdb tests/robot || :
 	docker-compose -f docker-compose-base.yml -f .ci/docker-compose.yml down --remove-orphans
 
-itest-crux:
+itest:
+	docker-compose -f docker-compose-base.yml -f .ci/docker-compose.yml build --parallel
+	docker-compose -f docker-compose-base.yml -f .ci/docker-compose.yml down
+	docker-compose -f docker-compose-base.yml -f .ci/docker-compose.yml  run --rm octopoes_integration
+
+rtest-crux:
 	docker-compose -f docker-compose-base.yml -f .ci/docker-compose-crux.yml up -d --build
 	sleep 1
 	robot -d reports --variablefile tests/robot/variables.py:crux tests/robot || :
