+++ octopoes/octopoes/xtdb/client.py
@@ -87,7 +87,7 @@ class XTDBHTTPClient:
         self._verify_response(res)
         return res.json()
 
-    def query(self, query: str, valid_time: Optional[datetime] = None) -> Dict:
+    def query(self, query: str, valid_time: Optional[datetime] = None) -> Union[List, Dict]:
         if valid_time is None:
             valid_time = datetime.now(timezone.utc)
         res = self._session.post(
