+++ octopoes/tests/conftest.py
@@ -4,12 +4,18 @@ from typing import Dict, List, Optional, Set
 from unittest.mock import Mock
 
 import pytest
+from octopoes.core.app import get_xtdb_client
+
+from octopoes.api.api import app
+from octopoes.api.router import settings
+from octopoes.config.settings import Settings, XTDBType
 
 from octopoes.models import OOI, EmptyScanProfile, Reference, ScanProfileBase
 from octopoes.models.path import Direction, Path
 from octopoes.models.types import DNSZone, Hostname, IPAddressV4, Network, ResolvedHostname
 from octopoes.repositories.ooi_repository import OOIRepository
 from octopoes.repositories.scan_profile_repository import ScanProfileRepository
+from octopoes.xtdb.client import XTDBHTTPClient
 
 
 @pytest.fixture
@@ -138,3 +144,23 @@ def resolved_hostname(hostname, ipaddressv4, ooi_repository, scan_profile_reposi
         scan_profile_repository,
         valid_time,
     )
+
+
+@pytest.fixture
+def xtdbtype_multinode():
+    def get_settings_override():
+        return Settings(xtdb_type=XTDBType.XTDB_MULTINODE)
+
+    app.dependency_overrides[settings] = get_settings_override
+    yield
+    app.dependency_overrides = {}
+
+
+@pytest.fixture
+def app_settings():
+    return Settings(xtdb_type=XTDBType.XTDB_MULTINODE)
+
+
+@pytest.fixture
+def xtdb_http_client(app_settings: Settings):
+    return get_xtdb_client(app_settings.xtdb_uri, "test", app_settings.xtdb_type)
